Refactor MixHistory component by removing the file and updating imports in App component. Move MixRecord type definition to a separate types file for better organization.
